backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moe-chat-backend"
version = "0.2.0"
description = "MoE-Chat 后端服务 - 基于混合专家模型的聊天应用"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "loguru>=0.7.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "alembic>=1.13.0",
    "jinja2>=3.0.0",
    "celery>=5.3.0",
    "flower>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "isort>=5.12.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "logs*", "uploads*"]

[tool.black]
line-length = 88  # Black 默认行长度
target-version = ["py312"]  # 指定 Python 版本
include = '\.pyi?$'  # 匹配 .py 和 .pyi 文件
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"  # 兼容 Black 的风格
line_length = 88
src_paths = ["src", "tests"]  # 指定扫描目录

[tool.flake8]
max-line-length = 88
max-complexity = 10
ignore = "E203,W503"
exclude = """
    .git,
    __pycache__,
    .venv/,
    build/,
    dist/,
    alembic/versions/,
    logs/,
    uploads/
"""
per-file-ignores = "__init__.py:F401"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[dependency-groups]
dev = [
    "types-python-jose>=3.5.0.20250531",
]
